+++ packages/nhsuk-frontend/src/nhsuk/components/character-count/character-count.mjs
@@ -1,9 +1,22 @@
-import { normaliseOptions } from '../../common/configuration/index.mjs'
+import { closestAttributeValue } from '../../common/closest-attribute-value.mjs'
+import {
+  normaliseOptions,
+  validateConfig
+} from '../../common/configuration/index.mjs'
+import { formatErrorMessage } from '../../common/index.mjs'
 import { ConfigurableComponent } from '../../configurable-component.mjs'
-import { ElementError } from '../../errors/index.mjs'
+import { ConfigError, ElementError } from '../../errors/index.mjs'
+import { I18n } from '../../i18n.mjs'
 
 /**
  * Character count component
+ *
+ * Tracks the number of characters or words in the `.nhsuk-js-character-count`
+ * `<textarea>` inside the element. Displays a message with the remaining number
+ * of characters/words available, or the number of characters/words in excess.
+ *
+ * You can configure the message to only appear after a certain percentage
+ * of the available characters/words has been entered.
  *
  * @augments ConfigurableComponent<CharacterCountConfig>
  */
@@ -21,14 +34,13 @@ export class CharacterCount extends ConfigurableComponent {
 
   /**
    * @param {Element | null} $root - HTML element to use for component
-   * @param {Record<string, never>} [config] - Not yet supported. Character count config
+   * @param {CharacterCountConfig} [config] - Character count config
    */
   constructor($root, config = {}) {
     super($root, config)
 
     const $textarea = this.$root.querySelector('.nhsuk-js-character-count')
     if (
-      !$textarea ||
       !(
         $textarea instanceof HTMLTextAreaElement ||
         $textarea instanceof HTMLInputElement
@@ -42,27 +54,47 @@ export class CharacterCount extends ConfigurableComponent {
       })
     }
 
+    // Check for valid config
+    const errors = validateConfig(CharacterCount.schema, this.config)
+    if (errors[0]) {
+      throw new ConfigError(formatErrorMessage(CharacterCount, errors[0]))
+    }
+
+    this.i18n = new I18n(this.config.i18n, {
+      // Read the fallback if necessary rather than have it set in the defaults
+      locale: closestAttributeValue(this.$root, 'lang')
+    })
+
+    // Determine the limit attribute (characters or words)
+    this.maxLength = this.config.maxwords ?? this.config.maxlength ?? Infinity
+
     this.$textarea = $textarea
 
-    const fallbackLimitMessageId = `${this.$textarea.id}-info`
-    const $fallbackLimitMessage = document.getElementById(
-      fallbackLimitMessageId
-    )
-
-    if (!$fallbackLimitMessage) {
+    const textareaDescriptionId = `${this.$textarea.id}-info`
+    const $textareaDescription = document.getElementById(textareaDescriptionId)
+    if (!$textareaDescription) {
       throw new ElementError({
         component: CharacterCount,
-        element: $fallbackLimitMessage,
-        identifier: `Count message (\`id="${fallbackLimitMessageId}"\`)`
+        element: $textareaDescription,
+        identifier: `Count message (\`id="${textareaDescriptionId}"\`)`
       })
     }
 
     // Pre-existing validation error rendered from server
     this.$errorMessage = this.$root.querySelector('.nhsuk-error-message')
 
-    // Move the fallback count message to be immediately after the textarea
+    // Inject a description for the textarea if none is present already
+    // for when the component was rendered with no maxlength, maxwords
+    // nor custom textareaDescriptionText
+    if (/^\s*$/.exec($textareaDescription.textContent)) {
+      $textareaDescription.textContent = this.i18n.t('textareaDescription', {
+        count: this.maxLength
+      })
+    }
+
+    // Move the textarea description to be immediately after the textarea
     // Kept for backwards compatibility
-    this.$textarea.insertAdjacentElement('afterend', $fallbackLimitMessage)
+    this.$textarea.insertAdjacentElement('afterend', $textareaDescription)
 
     // Create the *screen reader* specific live-updating counter
     // This doesn't need any styling classes, as it is never visible
@@ -71,28 +103,23 @@ export class CharacterCount extends ConfigurableComponent {
       'nhsuk-character-count__sr-status nhsuk-u-visually-hidden'
     $screenReaderCountMessage.setAttribute('aria-live', 'polite')
     this.$screenReaderCountMessage = $screenReaderCountMessage
-    $fallbackLimitMessage.insertAdjacentElement(
+    $textareaDescription.insertAdjacentElement(
       'afterend',
       $screenReaderCountMessage
     )
 
     // Create our live-updating counter element, copying the classes from the
-    // fallback element for backwards compatibility as these may have been configured
+    // textarea description for backwards compatibility as these may have been
+    // configured
     const $visibleCountMessage = document.createElement('div')
-    $visibleCountMessage.className = $fallbackLimitMessage.className
+    $visibleCountMessage.className = $textareaDescription.className
     $visibleCountMessage.classList.add('nhsuk-character-count__status')
     $visibleCountMessage.setAttribute('aria-hidden', 'true')
     this.$visibleCountMessage = $visibleCountMessage
-    $fallbackLimitMessage.insertAdjacentElement(
-      'afterend',
-      $visibleCountMessage
-    )
+    $textareaDescription.insertAdjacentElement('afterend', $visibleCountMessage)
 
-    // Hide the fallback limit message
-    $fallbackLimitMessage.classList.add('nhsuk-u-visually-hidden')
-
-    // Determine the limit attribute (characters or words)
-    this.maxLength = this.config.maxwords ?? this.config.maxlength ?? Infinity
+    // Hide the textarea description
+    $textareaDescription.classList.add('nhsuk-u-visually-hidden')
 
     // Remove hard limit if set
     this.$textarea.removeAttribute('maxlength')
@@ -111,23 +138,48 @@ export class CharacterCount extends ConfigurableComponent {
   }
 
   /**
-   * Counts characters or words in text
+   * Character count config override
    *
-   * @param {string} text
+   * To ensure data-attributes take complete precedence, even if they change
+   * the type of count, we need to reset the `maxlength` and `maxwords` from
+   * the JavaScript config.
+   *
+   * @param {Partial<CharacterCountConfig>} datasetConfig - Config specified by dataset
+   * @returns {Partial<CharacterCountConfig>} Config to override by dataset
    */
-  count(text) {
-    let length
-    if (this.config.maxwords) {
-      const tokens = text.match(/\S+/g) ?? [] // Matches consecutive non-whitespace chars
-      length = tokens.length
-    } else {
-      length = text.length
+  configOverride(datasetConfig) {
+    let configOverrides = {}
+    if ('maxwords' in datasetConfig || 'maxlength' in datasetConfig) {
+      configOverrides = {
+        maxlength: undefined,
+        maxwords: undefined
+      }
     }
-    return length
+
+    return configOverrides
   }
 
   /**
-   * Bind input propertychange to the elements and update based on the change
+   * Count the number of characters (or words, if `config.maxwords` is set)
+   * in the given text
+   *
+   * @param {string} text - The text to count the characters of
+   * @returns {number} the number of characters (or words) in the text
+   */
+  count(text) {
+    if (this.config.maxwords) {
+      const tokens = text.match(/\S+/g) ?? [] // Matches consecutive non-whitespace chars
+      return tokens.length
+    }
+
+    return text.length
+  }
+
+  /**
+   * Bind change events
+   *
+   * Set up event listeners on the $textarea so that the count messages update
+   * when the user types.
    */
   bindChangeEvents() {
     this.$textarea.addEventListener('keyup', () => this.handleKeyUp())
@@ -138,9 +190,7 @@ export class CharacterCount extends ConfigurableComponent {
   }
 
   /**
-   * Speech recognition software such as Dragon NaturallySpeaking will modify the
-   * fields by directly changing its `value`. These changes don't trigger events
-   * in JavaScript, so we need to poll to handle when and if they occur
+   * Update count message if textarea value has changed
    */
   checkIfValueChanged() {
     if (this.$textarea.value !== this.lastInputValue) {
@@ -150,6 +200,8 @@ export class CharacterCount extends ConfigurableComponent {
   }
 
   /**
+   * Update count message
+   *
    * Helper function to update both the visible and screen reader-specific
    * counters simultaneously (e.g. on init)
    */
@@ -159,14 +211,14 @@ export class CharacterCount extends ConfigurableComponent {
   }
 
   /**
-   * Update visible counter
+   * Update visible count message
    */
   updateVisibleCountMessage() {
     const remainingNumber = this.maxLength - this.count(this.$textarea.value)
     const isError = remainingNumber < 0
 
-    // If input is over the threshold, remove the disabled class which renders the
-    // counter invisible.
+    // If input is over the threshold, remove the disabled class which renders
+    // the counter invisible.
     this.$visibleCountMessage.classList.toggle(
       'nhsuk-character-count__message--disabled',
       !this.isOverThreshold()
@@ -183,11 +235,11 @@ export class CharacterCount extends ConfigurableComponent {
     this.$visibleCountMessage.classList.toggle('nhsuk-hint', !isError)
 
     // Update message
-    this.$visibleCountMessage.innerHTML = this.formattedUpdateMessage()
+    this.$visibleCountMessage.textContent = this.formattedUpdateMessage()
   }
 
   /**
-   * Update screen reader-specific counter
+   * Update screen reader count message
    */
   updateScreenReaderCountMessage() {
     // If over the threshold, remove the aria-hidden attribute, allowing screen
@@ -199,27 +251,39 @@ export class CharacterCount extends ConfigurableComponent {
     }
 
     // Update message
-    this.$screenReaderCountMessage.innerHTML = this.formattedUpdateMessage()
+    this.$screenReaderCountMessage.textContent = this.formattedUpdateMessage()
   }
 
   /**
-   * Format update message
+   * Get count message
+   *
+   * @returns {string} Status message
    */
   formattedUpdateMessage() {
     const remainingNumber = this.maxLength - this.count(this.$textarea.value)
+    const countType = this.config.maxwords ? 'words' : 'characters'
+    return this.formatCountMessage(remainingNumber, countType)
+  }
 
-    let charVerb = 'remaining'
-    let charNoun = 'character'
-    let displayNumber = remainingNumber
-    if (this.config.maxwords) {
-      charNoun = 'word'
+  /**
+   * Formats the message shown to users according to what's counted
+   * and how many remain
+   *
+   * @param {number} remainingNumber - The number of words/characaters remaining
+   * @param {string} countType - "words" or "characters"
+   * @returns {string} Status message
+   */
+  formatCountMessage(remainingNumber, countType) {
+    if (remainingNumber === 0) {
+      return this.i18n.t(`${countType}AtLimit`)
     }
-    charNoun += remainingNumber === -1 || remainingNumber === 1 ? '' : 's'
 
-    charVerb = remainingNumber < 0 ? 'too many' : 'remaining'
-    displayNumber = Math.abs(remainingNumber)
+    const translationKeySuffix =
+      remainingNumber < 0 ? 'OverLimit' : 'UnderLimit'
 
-    return `You have ${displayNumber} ${charNoun} ${charVerb}`
+    return this.i18n.t(`${countType}${translationKeySuffix}`, {
+      count: Math.abs(remainingNumber)
+    })
   }
 
   /**
@@ -227,15 +291,22 @@ export class CharacterCount extends ConfigurableComponent {
    *
    * Checks whether the value is over the configured threshold for the input.
    * If there is no configured threshold, it is set to 0 and this function will
-   * always return true
+   * always return true.
+   *
+   * @returns {boolean} true if the current count is over the config.threshold
+   *   (or no threshold is set)
    */
   isOverThreshold() {
+    // No threshold means we're always above threshold so save some computation
+    if (!this.config.threshold) {
+      return true
+    }
+
     // Determine the remaining number of characters/words
     const currentLength = this.count(this.$textarea.value)
+    const maxLength = this.maxLength
 
-    // Set threshold if presented in config
-    const thresholdPercent = this.config.threshold ?? 0
-    const thresholdValue = (this.maxLength * thresholdPercent) / 100
+    const thresholdValue = (maxLength * this.config.threshold) / 100
 
     return thresholdValue <= currentLength
   }
@@ -253,12 +324,18 @@ export class CharacterCount extends ConfigurableComponent {
 
   /**
    * Handle focus event
+   *
+   * Speech recognition software such as Dragon NaturallySpeaking will modify
+   * the fields by directly changing its `value`. These changes don't trigger
+   * events in JavaScript, so we need to poll to handle when and if they occur.
+   *
+   * Once the keyup event hasn't been detected for at least 1000 ms (1s), check
+   * if the textarea value has changed and update the count message if it has.
+   *
+   * This is so that the update triggered by the manual comparison doesn't
+   * conflict with debounced KeyboardEvent updates.
    */
   handleFocus() {
-    // If the field is focused, and a keyup event hasn't been detected for at
-    // least 1000 ms (1 second), then run the manual change check.
-    // This is so that the update triggered by the manual comparison doesn't
-    // conflict with debounced KeyboardEvent updates.
     this.valueChecker = window.setInterval(() => {
       if (
         !this.lastInputTimestamp ||
@@ -294,29 +371,65 @@ export class CharacterCount extends ConfigurableComponent {
    * @type {CharacterCountConfig}
    */
   static defaults = Object.freeze({
-    threshold: 0
+    threshold: 0,
+    i18n: {
+      // Characters
+      charactersUnderLimit: {
+        one: 'You have %{count} character remaining',
+        other: 'You have %{count} characters remaining'
+      },
+      charactersAtLimit: 'You have 0 characters remaining',
+      charactersOverLimit: {
+        one: 'You have %{count} character too many',
+        other: 'You have %{count} characters too many'
+      },
+      // Words
+      wordsUnderLimit: {
+        one: 'You have %{count} word remaining',
+        other: 'You have %{count} words remaining'
+      },
+      wordsAtLimit: 'You have 0 words remaining',
+      wordsOverLimit: {
+        one: 'You have %{count} word too many',
+        other: 'You have %{count} words too many'
+      },
+      textareaDescription: {
+        other: ''
+      }
+    }
   })
 
   /**
-   * Character config schema
+   * Character count config schema
    *
    * @constant
    * @satisfies {Schema<CharacterCountConfig>}
    */
   static schema = Object.freeze({
     properties: {
+      i18n: { type: 'object' },
       maxwords: { type: 'number' },
       maxlength: { type: 'number' },
       threshold: { type: 'number' }
-    }
+    },
+    anyOf: [
+      {
+        required: ['maxwords'],
+        errorMessage: 'Either "maxlength" or "maxwords" must be provided'
+      },
+      {
+        required: ['maxlength'],
+        errorMessage: 'Either "maxlength" or "maxwords" must be provided'
+      }
+    ]
   })
 }
 
 /**
  * Initialise character count component
  *
- * @deprecated Use {@link createAll | `createAll(CharacterCount)`} instead.
- * @param {InitOptions} [options]
+ * @deprecated Use {@link createAll | `createAll(CharacterCount, options)`} instead.
+ * @param {InitOptions & CharacterCountConfig} [options]
  */
 export function initCharacterCounts(options) {
   const { scope: $scope } = normaliseOptions(options)
@@ -326,7 +439,7 @@ export function initCharacterCounts(options) {
   )
 
   $characterCounts?.forEach(($root) => {
-    new CharacterCount($root)
+    new CharacterCount($root, options)
   })
 }
 
@@ -342,9 +455,55 @@ export function initCharacterCounts(options) {
  * @property {number} [threshold=0] - The percentage value of the limit at
  *   which point the count message is displayed. If this attribute is set, the
  *   count message will be hidden by default.
+ * @property {CharacterCountTranslations} [i18n=CharacterCount.defaults.i18n] - Character count translations
  */
 
 /**
+ * Character count translations
+ *
+ * @see {@link CharacterCount.defaults.i18n}
+ * @see {@link https://github.com/nhsuk/nhsuk-frontend/blob/main/docs/configuration/localisation.md}
+ * @typedef {object} CharacterCountTranslations
+ *
+ * Messages shown to users as they type. It provides feedback on how many words
+ * or characters they have remaining or if they are over the limit. This also
+ * includes a message used as an accessible description for the textarea.
+ * @property {TranslationPluralForms} [charactersUnderLimit] - Message displayed
+ *   when the number of characters is under the configured maximum, `maxlength`.
+ *   This message is displayed visually and through assistive technologies. The
+ *   component will replace the `%{count}` placeholder with the number of
+ *   remaining characters. This is a pluralised list of messages.
+ * @property {string} [charactersAtLimit] - Message displayed when the number of
+ *   characters reaches the configured maximum, `maxlength`. This message is
+ *   displayed visually and through assistive technologies.
+ * @property {TranslationPluralForms} [charactersOverLimit] - Message displayed
+ *   when the number of characters is over the configured maximum, `maxlength`.
+ *   This message is displayed visually and through assistive technologies. The
+ *   component will replace the `%{count}` placeholder with the number of
+ *   remaining characters. This is a pluralised list of messages.
+ * @property {TranslationPluralForms} [wordsUnderLimit] - Message displayed when
+ *   the number of words is under the configured maximum, `maxlength`. This
+ *   message is displayed visually and through assistive technologies. The
+ *   component will replace the `%{count}` placeholder with the number of
+ *   remaining words. This is a pluralised list of messages.
+ * @property {string} [wordsAtLimit] - Message displayed when the number of
+ *   words reaches the configured maximum, `maxlength`. This message is
+ *   displayed visually and through assistive technologies.
+ * @property {TranslationPluralForms} [wordsOverLimit] - Message displayed when
+ *   the number of words is over the configured maximum, `maxlength`. This
+ *   message is displayed visually and through assistive technologies. The
+ *   component will replace the `%{count}` placeholder with the number of
+ *   remaining words. This is a pluralised list of messages.
+ * @property {TranslationPluralForms} [textareaDescription] - Message made
+ *   available to assistive technologies, if none is already present in the
+ *   HTML, to describe that the component accepts only a limited amount of
+ *   content. It is visible on the page when JavaScript is unavailable. The
+ *   component will replace the `%{count}` placeholder with the value of the
+ *   `maxlength` or `maxwords` parameter.
+ */
+
+/**
+ * @import { TranslationPluralForms } from '../../i18n.mjs'
  * @import { createAll, InitOptions } from '../../index.mjs'
  * @import { Schema } from '../../common/configuration/index.mjs'
  */