+++ CHANGELOG.md
@@ -4,21 +4,87 @@
 
 ### :new: **New features**
 
+#### Localise character count component
+
+You can now translate the text used by the [character count](https://service-manual.nhs.uk/design-system/components/character-count) component to:
+
+- show when the maximum number of characters or words is reached
+- show the number of characters or words over or under the allowed maximum
+- update the textarea description if JavaScript is not available
+- announce to screen readers when the textarea is focused
+
+The Nunjucks macro accepts new options so you can customise each message. You can:
+
+- Use `charactersAtLimitText` or `wordsAtLimitText` to provide the text that shows when users have reached the limit.
+- Use `charactersUnderLimitText` or `wordsUnderLimitText` to provide the text that shows when users are under the limit. The component will pluralise the message according to the configured locale and the number of characters or words remaining.
+- Use `charactersOverLimitText` or `wordsOverLimitText` to provide the text that shows when users are over the limit. The component will pluralise the message according to the configured locale and the number of characters or words remaining.
+- Use `textareaDescriptionText` to provide the textarea description for assistive technologies. It is visible on the page when JavaScript is unavailable.
+
+The component will replace `%{count}` with the number of characters over or under the limit.
+
+If you're not using Nunjucks macros, you can use data-\* attributes to provide these translations. Within the attribute value, any [quotation marks or other characters reserved by HTML](https://developer.mozilla.org/en-US/docs/Glossary/Character_reference) needs to be converted into their HTML entity equivalents.
+
+You can:
+
+- use `data-i18n.characters-at-limit` or `data-i18n.words-at-limit` for when users are at the limit
+- configure the text that informs the end user they are under the character or word limit, by using `data-i18n.characters-under-limit.{other,many,few,two,one,zero}` or `data-i18n.words-under-limit.{other,many,few,two,one,zero}`, with one suffix for each plural form required by your locale
+- configure the text that informs the end user they are over the character or word limit, by using `data-i18n.characters-over-limit.{other,many,few,two,one,zero}` or `data-i18n.words-over-limit.{other,many,few,two,one,zero}`, with one suffix for each plural form required by your locale
+- configure the description provided to assistive technologies when users focus the input, by using `data-i18n.textarea-description.{other,many,few,two,one,zero}` to provide the text to set as the description
+
+You can also provide these messages using a JavaScript configuration object when creating an instance of the component or initialising all components. See [our guidance on localising NHS.UK frontend](/docs/configuration/localisation.md) for how to do this.
+
+This was added in [pull request #1565: Add character count localisation](https://github.com/nhsuk/nhsuk-frontend/pull/1565).
+
+#### Smaller versions of radio buttons and checkboxes
+
+You can now use smaller versions of the [radios](https://service-manual.nhs.uk/design-system/components/radios) and [checkboxes](https://service-manual.nhs.uk/design-system/components/checkboxes) components by adding the `nhsuk-radios--small` or `nhsuk-checkboxes--small` class.
+
+This was added in [pull request #1567: Uplift GOV.UK Frontend form field components](https://github.com/nhsuk/nhsuk-frontend/pull/1567).
+
+#### Updated Nunjucks macro options for components
+
+For consistency with other components, we’ve added new Nunjucks macro options:
+
+- Back link `visuallyHiddenText` option
+- Breadcrumbs nested `backLink` component option
+- Contents list `landmarkLabel` and `visuallyHiddenTitle` options
+- Do and Don't list `prefixText` option, with nested item `text` and `html` options
+- Fieldset `html` and `role` options
+- Header navigation `toggleMenuText` and `toggleMenuVisuallyHiddenText` options
+- Inset text `visuallyHiddenText` option
+- Pagination `previous`, `next` and `landmarkLabel` options
+- Radios item `disabled` and `attributes` options
+- Text input `autocapitalize` and `disabled` options, with prefix and suffix nested `text`, `html`, `classes` and `attributes` options
+- Textarea `disabled` and `spellcheck` options
+- Warning callout `visuallyHiddenText` option
+
+Visit the [design system in the NHS digital service manual](https://service-manual.nhs.uk/design-system/components) to see Nunjucks options for each component.
+
+This was added in [pull request #1567: Uplift GOV.UK Frontend form field components](https://github.com/nhsuk/nhsuk-frontend/pull/1567).
+
+#### Updated appearance of disabled form controls
+
+We’ve updated the disabled state of radio buttons, checkboxes, text inputs, textareas and select menus so it is consistent across browsers and devices.
+
+Disabled form controls appear at 50% opacity and with an alternative cursor appearance when hovered over.
+
+This was added in [pull request #1567: Uplift GOV.UK Frontend form field components](https://github.com/nhsuk/nhsuk-frontend/pull/1567).
+
 #### Insert custom HTML into component form group wrappers
 
 You can now insert custom HTML into form group wrappers for all components with form fields.
 
 ```njk
-input({
+{{ input({
   formGroup: {
     beforeInput: {
       html: "example"
     },
     afterInput: {
       html: "example"
-    },
+    }
   }
-})
+}) }}
 ```
 
 This was added in [pull request #1561: Add `beforeInput(s)` and `beforeInput(s)` options to form groups](https://github.com/nhsuk/nhsuk-frontend/pull/1561).
@@ -66,11 +132,74 @@ The previous names are deprecated and will be removed in a future release.
 
 This change was introduced in [pull request #1594: Always set input `:focus` box-shadow colour](https://github.com/nhsuk/nhsuk-frontend/pull/1594).
 
+### :recycle: **Changes**
+
+#### Do and Don't list Nunjucks options
+
+For consistency with other components, the do and don't list Nunjucks options have changed. The previous names are deprecated and will be removed in a future release.
+
+If you're using the `list` Nunjucks macro in your service, you should update the nested `items` option, using `text` or `html` instead of `item`.
+
+```patch
+  {{ list({
+    title: "Do",
+    type: "tick",
+    items: [
+      {
+-       item: "cover blisters with a soft plaster or padded dressing"
++       text: "cover blisters with a soft plaster or padded dressing"
+      },
+      {
+-       item: "wash your hands before touching a burst blister"
++       text: "wash your hands before touching a burst blister"
+      },
+      {
+-       item: "allow the fluid in a burst blister to drain before covering it with a plaster or dressing"
++       text: "allow the fluid in a burst blister to drain before covering it with a plaster or dressing"
+      }
+    ]
+  }) }}
+```
+
+This change was introduced in [pull request #1620: Make all Nunjucks components text configurable](https://github.com/nhsuk/nhsuk-frontend/pull/1620).
+
+#### Pagination Nunjucks options
+
+For consistency with other components, the pagination Nunjucks options have changed. The previous names are deprecated and will be removed in a future release.
+
+If you're using the `pagination` Nunjucks macro in your service, you should:
+
+- replace the `previousUrl` option with the nested `previous.href` option
+- replace the `previousPage` option with the nested `previous.labelText` option
+- replace the `nextUrl` option with the nested `next.href` option
+- replace the `nextPage` option with the nested `next.labelText` option
+
+```patch
+  {{ pagination({
+-   previousPage: "Treatments",
+-   previousUrl: "/section/treatments",
++   previous: {
++     labelText: "Treatments",
++     href: "/section/treatments"
++   },
+-   nextPage: "Symptoms",
+-   nextUrl: "/section/symptoms"
++   next: {
++     labelText: "Symptoms",
++     href: "/section/symptoms"
++   }
+  }) }}
+```
+
+This change was introduced in [pull request #1620: Make all Nunjucks components text configurable](https://github.com/nhsuk/nhsuk-frontend/pull/1620).
+
 ### :wrench: **Fixes**
 
 - [#1593: Fix Sass deprecated `core/settings/colours` partial](https://github.com/nhsuk/nhsuk-frontend/pull/1593)
 - [#1594: Always set input `:focus` box-shadow colour](https://github.com/nhsuk/nhsuk-frontend/pull/1594)
 - [#1599: Make border colour on `nhsuk-panel` mixin optional](https://github.com/nhsuk/nhsuk-frontend/pull/1599)
+- [#1601: Fix accessible name for linked logo in header component](https://github.com/nhsuk/nhsuk-frontend/pull/1601)
+- [#1612: Fix styling of action link when used in emergency cards ](https://github.com/nhsuk/nhsuk-frontend/pull/1612)
 
 ## 10.0.0 - 26 August 2025